lib.attention: use local attention instead of global...
Since global soft alignment tends to 'jump' back to
earlier occurences of the same character, which cannot
be remedied by favouring monotonic alternatives in the
decoder (because the jumps are not limited to alternatives),
and which becomes more severe with longer sequences,
we need a local soft alignment mechanism.

We have no need for elaborate position proposal models
though: OCR errors which can be recovered at all are
always localized already. Hence, all we need is soft
alignment of a fixed window around the position of the
current target length in the source (monotonic alignment).

The base implementation only provides global attention.
This is a provisional implementation of local attention.